+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes a bug where Go generation would map the wrong http core method for primitive types #573
 - Fixes a bug where Go generation would have unused imports making build fail #573
 - Fixes a bug where missing type definitions would make Ruby generation fail #573
+- Fixes a bug where Go generation would miss the module symbol for inherited constructors #573
 
 ## [0.0.8] - 2021-08-25